aiModel/huggingSpace/requirements.txt
gradio>=4.0.0
rembg>=2.0.55
pillow>=10.1.0
onnxruntime-gpu>=1.16.0
